Evict tags from core
