+++ apecs/apecs.cabal
@@ -33,6 +33,7 @@ library
     Apecs.Experimental.Util
     Apecs.Stores
     Apecs.System
+    Apecs.Tags
     Apecs.TH
     Apecs.TH.Tags
     Apecs.Util
