+++ apecs/src/Apecs/Core.hs
@@ -81,12 +81,3 @@ type Get     w m c = (Has w m c, ExplGet     m (Storage c))
 type Set     w m c = (Has w m c, ExplSet     m (Storage c))
 type Members w m c = (Has w m c, ExplMembers m (Storage c))
 type Destroy w m c = (Has w m c, ExplDestroy m (Storage c))
-
--- | The type of tags for a world, e.g. @WTag MyWorld = MyWorldTag@.
---   Standalone so that multiple @HasTags w m@ instances share one equation.
-type family WTag w
-
--- | @HasTags w m@ means that world @w@ has a tag system generated by @makeTaggedComponents@.
---   Provides a way to query component tags for entities, dispatching on the world type @w@.
-class Monad m => HasTags w m where
-  entityTags :: Entity -> SystemT w m [WTag w]
