+++ apecs/test/Main.hs
@@ -34,6 +34,7 @@ import           Apecs.Experimental.Children
 import           Apecs.Experimental.Reactive
 import           Apecs.Experimental.Stores
 import           Apecs.Stores
+import           Apecs.Tags
 import           Apecs.TH
 import           Apecs.TH.Tags
 import           Apecs.Util
